[GovWayCore, InstallerVersioneBinaria, GovWayConsole, GovWayMonitor, ApiRsConfigurazione, ApiRsMonitoraggio] Aggiunto supporto per gestire keystore firmati.
[GovWayCore]
Aggiunto supporto per collezionare una serie di variabili Java che potranno essere riferite in qualsiasi file di proprietà di GovWay presente nella 'directory-lavoro' tramite la sintassi '${nomeVar}'.
Le variabili potranno inoltre essere accedute all’interno delle varie configurazioni di GovWay come, ad esempio, tramite la sintassi 'java:NAME' o 'envj:NAME'.
La definizione può essere attuata all'interno di due file differenti:
- govway.map.properties: le variabili definite in questo file verranno caricate all'avvio di ogni applicazione GovWay (es. runtime, console, batch ecc.)
- govway.secrets.properties: i valori delle variabili definite nel file saranno cifrati e GovWay si occuperà di decifrarli prima del loro caricamento nel sistema.

[GovWayCore, GovWayConsole, ApiRsConfigurazione]
Le informazioni confidenziali presenti nella base dati possono adesso essere cifrate tramite le utility 'Bring Your Own Key (BYOK)'.
La definizione della master key e le modalità di cifratura e decifratura sono specificate all'interno di un file di configurazione presente nella 'directory-lavoro' con il nome 'byok.properties'.
La console di gestione è stata rivista per supportare i nuovi campi contenenti informazioni confidenziali.

[GovWayConsole]
Il corretto export dei plugin per l'autorizzazione dei contenuti non veniva eseguito correttamente, sia durante l'erogazione che durante la fruizione.

[GovWayConsole]
(https://github.com/link-it/govway/issues/159)
Nella configurazione del connettore è adesso possibile specificare il metodo di autenticazione 'api-key'.

[GovWayCore, GovWayConsole]
La verifica dei certificati include adesso anche la validazione dei keystore riferiti nella configurazione della sicurezza dei messaggi.
Inoltre, è stata aggiunta la possibilità di disabilitare la 'Compliance BSP 1.1' nella validazione di un messaggio contenente WS-Security Username Token.

[GovWayCore]
Aggiunta la possibilità di utilizzare policy OCSP nei connettori HTTPS riferiti nelle token policy di validazione e negoziazione.
Risolto inoltre un problema che si presentava selezionando un'autenticazione HTTPS client sulla funzionalità specifica di introspection o userinfo, ma non si abilitava l'endpoint HTTPS nella sezione token.
In questo scenario, la gestione personalizzata dei keystore utilizzati per la connessione HTTPS non veniva attivata.

[GovWayConsole]
Aggiunta la possibilità di svuotare la cache e verificare i certificati in qualsiasi maschera di configurazione di erogazione o fruizione di API, oltre che nella scheda di dettaglio.

[GovWayConsole]
Corretto il problema presente nella funzionalità di export per il profilo SPCoop che si verificava quando veniva selezionato un soggetto multi-tenant tra quelli disponibili.
Il problema si presentava durante l'export di una API o di una fruizione che riferiva una API il cui soggetto era differente da quello selezionato; all'interno dell'archivio zip non veniva inserita l'API.

+++ core/src/org/openspcoop2/pdd/core/dynamic/test/DynamicTest.java
@@ -1948,6 +1948,58 @@ public class DynamicTest {
 			throw new Exception("Expected null value, found '"+value+"'");
 		}
 		
+		expr = prefix + "{envj:pCustomJava1}";
+		DynamicUtils.validate("testEnvJava", expr, addPrefixError);
+		value = DynamicUtils.convertDynamicPropertyValue("testJava", expr, dynamicMap, pddContext);
+		System.out.println("testEnvJava: "+value+"\n\n");
+		expected = "v1java";
+		if(!expected.equals(value)) {
+			throw new Exception("Expected value '"+expected+"', found '"+value+"'");
+		}
+		
+		expr = prefix+"{envj:javaNotFound}";
+		DynamicUtils.validate("testEnvJavaNotFound", expr, addPrefixError);
+		value = DynamicUtils.convertDynamicPropertyValue("testEnvJavaNotFound", expr, dynamicMap, pddContext);
+		System.out.println("testEnvJavaNotFound: "+value);
+		expected = "";
+		if(!expected.equals(value)) {
+			throw new Exception("Expected value '"+expected+"', found '"+value+"'");
+		}
+		
+		expr = prefixOptional+"{envj:javaNotFound}";
+		DynamicUtils.validate("optional testEnvJavaNotFound", expr, addPrefixError);
+		value = DynamicUtils.convertDynamicPropertyValue("optional testEnvJavaNotFound", expr, dynamicMap, pddContext);
+		System.out.println("optional testEnvJavaNotFound: "+value+"\n\n");
+		if(value!=null) {
+			throw new Exception("Expected null value, found '"+value+"'");
+		}
+		
+		expr = prefix + "{envj:HOSTNAME}";
+		DynamicUtils.validate("testENVJava", expr, addPrefixError);
+		value = DynamicUtils.convertDynamicPropertyValue("testENVJava", expr, dynamicMap, pddContext);
+		System.out.println("testENVJava: "+value+"\n\n");
+		expected = System.getenv("HOSTNAME");
+		if(!expected.equals(value)) {
+			throw new Exception("Expected value '"+expected+"', found '"+value+"'");
+		}
+		
+		expr = prefix+"{envj:envNotFound}";
+		DynamicUtils.validate("testENVJavaNotFound", expr, addPrefixError);
+		value = DynamicUtils.convertDynamicPropertyValue("testENVJavaNotFound", expr, dynamicMap, pddContext);
+		System.out.println("testENVJavaNotFound: "+value);
+		expected = "";
+		if(!expected.equals(value)) {
+			throw new Exception("Expected value '"+expected+"', found '"+value+"'");
+		}
+		
+		expr = prefixOptional+"{envj:envNotFound}";
+		DynamicUtils.validate("optional testEVNJavaNotFound", expr, addPrefixError);
+		value = DynamicUtils.convertDynamicPropertyValue("optional testEVNJavaNotFound", expr, dynamicMap, pddContext);
+		System.out.println("optional testEVNJavaNotFound: "+value+"\n\n");
+		if(value!=null) {
+			throw new Exception("Expected null value, found '"+value+"'");
+		}
+		
 		}finally {
 			if(fTmpConfig!=null) {
 				fTmpConfig.delete();
